references: add default measure unit to the score wizard
